requirements.txt
mockito
PyYAML
gitpython
